+++ teamBootstrap/src/main/resources/templates/system/qualityInspectionList.html
@@ -140,7 +140,7 @@ layout:decorate="@{layout/default_system}">
                      	
                     </td>
                    <td>  <button class="btn btn-warning btn-xs" type="button" onclick="location.href='/addStandardTable'"><i class="fa fa-pencil"></i>기준표등록</button></td>
-                   <td>  <a th:href="@{/modifyQualityInspection(qualityInspectionCode=${l.quality_inspection_code})}" class="btn btn-primary btn-xs"> <i class="fa fa-pencil"></i>수정</a></td>
+                   <td>  <a th:href="@{/system/modifyQualityInspection(qualityInspectionCode=${l.quality_inspection_code})}" class="btn btn-primary btn-xs"> <i class="fa fa-pencil"></i>수정</a></td>
                    
                   </tr>
